docs/saas/arr-rebuild/source/backend/archive_unpacked/saasot/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ecdsa==0.18.0
enum-compat==0.0.3
et-xmlfile==1.1.0
future==0.18.3
idna==3.4
intuit-oauth==1.2.4
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.1
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-quickbooks==0.9.3
python3-openid==3.2.0
pytz==2023.3
rauth==0.7.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.19.1
six==1.16.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.2
